pyproject.toml
[project]
name = "itmo-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.11",
    "bs4>=0.0.2",
    "duckduckgo-search>=7.3.0",
    "fastapi>=0.115.8",
    "ipykernel>=6.29.5",
    "openai>=1.60.2",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "tavily-python>=0.5.0",
    "uvicorn>=0.34.0",
]

packages = [
    { include = "app" },
    { include = "llm" },
]
